# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Attendance Management API
  version: 1.0.0
  description: >
    Sistem absensi sekolah berbasis Node.js, Express, Sequelize, RBAC, dan semester.
    Mendukung absensi harian, absensi per jam, validasi, rekap, data master,
    manajemen user, scheduler otomatis, export, file bukti, serta logging aktivitas.
servers:
  - url: http://localhost:3000
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Jurusan:
      type: object
      properties:
        id: { type: integer }
        kode_jurusan: { type: string }
        nama_jurusan: { type: string }
...
paths:
...
  /api/jurusan:
    get:
      tags: [Jurusan]
      summary: List jurusan
      security:
        - BearerAuth: []
    post:
      tags: [Jurusan]
      summary: Create jurusan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Jurusan' }
  /api/jurusan/{id}:
    get:
      tags: [Jurusan]
      summary: Detail jurusan
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          schema: { type: integer }
          required: true
    put:
      tags: [Jurusan]
      summary: Update jurusan
      security:
        - BearerAuth: []
    delete:
      tags: [Jurusan]
      summary: Delete jurusan
      security:
        - BearerAuth: []


  ##############################################
  # KELAS
  ##############################################
  /api/kelas:
    get:
      tags: [Kelas]
      summary: List kelas
      security:
        - BearerAuth: []
    post:
      tags: [Kelas]
      summary: Create kelas
      security:
        - BearerAuth: []
  /api/kelas/{id}:
    get:
      tags: [Kelas]
      summary: Detail kelas
      security:
        - BearerAuth: []
    put:
      tags: [Kelas]
      summary: Update kelas
      security:
        - BearerAuth: []
    delete:
      tags: [Kelas]
      summary: Delete kelas
      security:
        - BearerAuth: []


  ##############################################
  # MAPEL
  ##############################################
  /api/mapel:
    get:
      tags: [Mapel]
      summary: List mapel
      security:
        - BearerAuth: []
    post:
      tags: [Mapel]
      summary: Create mapel
      security:
        - BearerAuth: []
...
  /api/absensi/input-harian:
    post:
      tags: [Absensi]
      summary: Input absensi harian
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                student_id: { type: integer }
                tanggal: { type: string, format: date }
                status:
                  type: string
                  enum: [hadir, sakit, izin, tanpa_keterangan]
                keterangan: { type: string }
                bukti:
                  type: string
                  format: binary
      responses:
        201:
          description: Absensi harian berhasil
  /api/absensi/input-jam:
    post:
      tags: [Absensi]
      summary: Input absensi per jam
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                student_id: { type: integer }
                tanggal: { type: string, format: date }
                jam_ke: { type: integer }
                status:
                  type: string
                  enum: [hadir, sakit, izin, tanpa_keterangan]
                bukti:
                  type: string
                  format: binary
      responses:
        201:
          description: Absensi jam-ke berhasil
  /api/absensi/validation-queue:
    get:
      tags: [Absensi]
      summary: Queue absensi menunggu validasi
      security:
        - BearerAuth: []
  /api/absensi/validate:
    post:
      tags: [Absensi]
      summary: Validasi absensi oleh guru mapel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                absensi_id: { type: integer }
                status_validasi: { type: string }
  /api/absensi/rekap-harian:
    get:
      tags: [Absensi]
      summary: Rekap harian
      security:
        - BearerAuth: []
  /api/absensi/rekap-bulanan:
    get:
      tags: [Absensi]
      summary: Rekap bulanan
      security:
        - BearerAuth: []
  /api/absensi/rekap-kelas:
    get:
      tags: [Absensi]
      summary: Rekap per kelas
      security:
        - BearerAuth: []
  /api/absensi/ranking-siswa:
    get:
      tags: [Absensi]
      summary: Ranking siswa
      security:
        - BearerAuth: []
  /api/absensi/riwayat:
    get:
      tags: [Absensi]
      summary: Riwayat absensi siswa
      security:
        - BearerAuth: []
  /api/absensi/export/csv:
    get:
      tags: [Absensi]
      summary: Export CSV
      security:
        - BearerAuth: []
  /api/absensi/export/pdf:
    get:
      tags: [Absensi]
      summary: Export PDF
      security:
        - BearerAuth: []


  ##############################################
  # ACTIVITY LOG
  ##############################################
...